(PC-34814) feat(ui): GenericInfoPageWhite > GenericInfoPageWhiteLegacy
