+++ src/features/auth/pages/suspendedAccount/AccountReactivationSuccess/AccountReactivationSuccess.tsx
@@ -5,11 +5,11 @@ import { navigateToHomeConfig } from 'features/navigation/helpers/navigateToHome
 import QpiThanks from 'ui/animations/qpi_thanks.json'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { InternalTouchableLink } from 'ui/components/touchableLink/InternalTouchableLink'
-import { GenericInfoPageWhite } from 'ui/pages/GenericInfoPageWhite'
+import { GenericInfoPageWhiteLegacy } from 'ui/pages/GenericInfoPageWhiteLegacy'
 import { getSpacing, Spacer, TypoDS } from 'ui/theme'
 
 export const AccountReactivationSuccess = () => (
-  <GenericInfoPageWhite
+  <GenericInfoPageWhiteLegacy
     mobileBottomFlex={0.1}
     animation={QpiThanks}
     title="Ton compte a été réactivé">
@@ -23,7 +23,7 @@ export const AccountReactivationSuccess = () => (
         navigateTo={{ ...navigateToHomeConfig, params: { ...navigateToHomeConfig.params } }}
       />
     </ButtonContainer>
-  </GenericInfoPageWhite>
+  </GenericInfoPageWhiteLegacy>
 )
 
 const ButtonContainer = styled.View({
