+++ src/features/profile/pages/ConfirmChangeEmail/ConfirmChangeEmail.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
 
 import { ConfirmChangeEmailContent } from 'features/profile/pages/ConfirmChangeEmail/ConfirmChangeEmailContent'
-import { GenericInfoPageWhite } from 'ui/pages/GenericInfoPageWhite'
+import { GenericInfoPageWhiteLegacy } from 'ui/pages/GenericInfoPageWhiteLegacy'
 import { BicolorPhonePending } from 'ui/svg/icons/BicolorPhonePending'
 import { Spacer, TypoDS } from 'ui/theme'
 
 export function ConfirmChangeEmail() {
   return (
-    <GenericInfoPageWhite
+    <GenericInfoPageWhiteLegacy
       icon={BicolorPhonePending}
       titleComponent={TypoDS.Title3}
       title="Confirmes-tu la demande de changement d’e-mail&nbsp;?"
@@ -15,6 +15,6 @@ export function ConfirmChangeEmail() {
       mobileBottomFlex={0.3}>
       <Spacer.Column numberOfSpaces={40} />
       <ConfirmChangeEmailContent />
-    </GenericInfoPageWhite>
+    </GenericInfoPageWhiteLegacy>
   )
 }
