+++ src/features/profile/pages/DeleteProfile/SuspendAccountConfirmationWithoutAuthentication.tsx
@@ -13,7 +13,7 @@ import { ButtonTertiaryBlack } from 'ui/components/buttons/ButtonTertiaryBlack'
 import { SNACK_BAR_TIME_OUT, useSnackBarContext } from 'ui/components/snackBar/SnackBarContext'
 import { ExternalTouchableLink } from 'ui/components/touchableLink/ExternalTouchableLink'
 import { VerticalUl } from 'ui/components/Ul'
-import { GenericInfoPageWhite } from 'ui/pages/GenericInfoPageWhite'
+import { GenericInfoPageWhiteLegacy } from 'ui/pages/GenericInfoPageWhiteLegacy'
 import { BicolorUserError } from 'ui/svg/BicolorUserError'
 import { EmailFilled } from 'ui/svg/icons/EmailFilled'
 import { ExternalSiteFilled } from 'ui/svg/icons/ExternalSiteFilled'
@@ -42,7 +42,7 @@ export const SuspendAccountConfirmationWithoutAuthentication: FC = () => {
   })
 
   return (
-    <GenericInfoPageWhite
+    <GenericInfoPageWhiteLegacy
       headerGoBack
       titleComponent={TypoDS.Title3}
       title="Souhaites-tu suspendre ton compte pass&nbsp;Culture&nbsp;?"
@@ -87,7 +87,7 @@ export const SuspendAccountConfirmationWithoutAuthentication: FC = () => {
           externalNav={{ url: `mailto:${env.FRAUD_EMAIL_ADDRESS}` }}
         />
       </ButtonContainer>
-    </GenericInfoPageWhite>
+    </GenericInfoPageWhiteLegacy>
   )
 }
